Add FactoryFacade methode to create comparison node
