+++ projects/com.oracle.truffle.llvm.parser.factories/src/com/oracle/truffle/llvm/parser/factories/LLVMComparisonFactory.java
@@ -139,12 +139,110 @@ import com.oracle.truffle.llvm.nodes.impl.op.compare.LLVMI8CompareNodeFactory.LL
 import com.oracle.truffle.llvm.nodes.impl.op.compare.LLVMI8CompareNodeFactory.LLVMI8UltNodeGen;
 import com.oracle.truffle.llvm.nodes.impl.op.compare.LLVMNeqNodeGen;
 import com.oracle.truffle.llvm.parser.LLVMBaseType;
+import com.oracle.truffle.llvm.parser.base.model.enums.CompareOperator;
+import com.oracle.truffle.llvm.parser.base.model.types.Type;
+import com.oracle.truffle.llvm.parser.base.util.LLVMParserRuntime;
+import com.oracle.truffle.llvm.parser.base.util.LLVMTypeHelper;
 import com.oracle.truffle.llvm.parser.instructions.LLVMFloatComparisonType;
 import com.oracle.truffle.llvm.parser.instructions.LLVMIntegerComparisonType;
 
 public final class LLVMComparisonFactory {
 
     private LLVMComparisonFactory() {
+    }
+
+    public static LLVMExpressionNode toCompareNode(CompareOperator operator, Type type, LLVMExpressionNode lhs, LLVMExpressionNode rhs) {
+        if (LLVMTypeHelper.isVectorType(type.getLLVMBaseType())) {
+            throw new AssertionError("We need a LLVMParserRuntime when creating an vector compare node!");
+        }
+
+        return toCompareVectorNode(null, operator, type, lhs, rhs);
+    }
+
+    public static LLVMExpressionNode toCompareVectorNode(LLVMParserRuntime runtime, CompareOperator operator, Type type, LLVMExpressionNode lhs, LLVMExpressionNode rhs) {
+        final LLVMBaseType llvmtype = type.getLLVMBaseType();
+
+        switch (operator) {
+            case FP_FALSE:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.FALSE);
+            case FP_ORDERED_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_EQUALS);
+            case FP_ORDERED_GREATER_THAN:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_GREATER_THAN);
+            case FP_ORDERED_GREATER_OR_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_GREATER_EQUALS);
+            case FP_ORDERED_LESS_THAN:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_LESS_THAN);
+            case FP_ORDERED_LESS_OR_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_LESS_EQUALS);
+            case FP_ORDERED_NOT_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED_AND_NOT_EQUALS);
+
+            case FP_ORDERED:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.ORDERED);
+            case FP_UNORDERED:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED);
+            case FP_UNORDERED_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_EQUALS);
+            case FP_UNORDERED_GREATER_THAN:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_GREATER_THAN);
+            case FP_UNORDERED_GREATER_OR_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_GREATER_EQUALS);
+            case FP_UNORDERED_LESS_THAN:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_LESS_THAN);
+            case FP_UNORDERED_LESS_OR_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_LESS_EQUALS);
+            case FP_UNORDERED_NOT_EQUAL:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.UNORDERED_OR_NOT_EQUALS);
+            case FP_TRUE:
+                return LLVMComparisonFactory.createFloatComparison(lhs, rhs, llvmtype, LLVMFloatComparisonType.TRUE);
+            default:
+                break;
+        }
+
+        final LLVMIntegerComparisonType comparison;
+        switch (operator) {
+            case INT_EQUAL:
+                comparison = LLVMIntegerComparisonType.EQUALS;
+                break;
+            case INT_NOT_EQUAL:
+                comparison = LLVMIntegerComparisonType.NOT_EQUALS;
+                break;
+            case INT_UNSIGNED_GREATER_THAN:
+                comparison = LLVMIntegerComparisonType.UNSIGNED_GREATER_THAN;
+                break;
+            case INT_UNSIGNED_GREATER_OR_EQUAL:
+                comparison = LLVMIntegerComparisonType.UNSIGNED_GREATER_EQUALS;
+                break;
+            case INT_UNSIGNED_LESS_THAN:
+                comparison = LLVMIntegerComparisonType.UNSIGNED_LESS_THAN;
+                break;
+            case INT_UNSIGNED_LESS_OR_EQUAL:
+                comparison = LLVMIntegerComparisonType.UNSIGNED_LESS_EQUALS;
+                break;
+            case INT_SIGNED_GREATER_THAN:
+                comparison = LLVMIntegerComparisonType.SIGNED_GREATER_THAN;
+                break;
+            case INT_SIGNED_GREATER_OR_EQUAL:
+                comparison = LLVMIntegerComparisonType.SIGNED_GREATER_EQUALS;
+                break;
+            case INT_SIGNED_LESS_THAN:
+                comparison = LLVMIntegerComparisonType.SIGNED_LESS_THAN;
+                break;
+            case INT_SIGNED_LESS_OR_EQUAL:
+                comparison = LLVMIntegerComparisonType.SIGNED_LESS_EQUALS;
+                break;
+
+            default:
+                throw new RuntimeException("Missed a compare operator");
+        }
+
+        if (LLVMTypeHelper.isVectorType(llvmtype)) {
+            final LLVMAddressNode target = (LLVMAddressNode) runtime.allocateVectorResult(type);
+            return LLVMComparisonFactory.createVectorComparison(target, lhs, rhs, llvmtype, comparison);
+        } else {
+            return LLVMComparisonFactory.createIntegerComparison(lhs, rhs, llvmtype, comparison);
+        }
     }
 
     public static LLVMI1Node createIntegerComparison(LLVMExpressionNode left, LLVMExpressionNode right, LLVMBaseType llvmType, LLVMIntegerComparisonType condition) {
